+++ src/CoordGeodetic.cpp
@@ -15,4 +15,4 @@
  */
 
 
-#include "CoordGeodetic.h"
+#include "csgp4/CoordGeodetic.h"
