+++ src/Globals.cpp
@@ -15,4 +15,4 @@
  */
 
 
-#include "Globals.h"
+#include "csgp4/Globals.h"
