+++ src/SatelliteException.cpp
@@ -14,4 +14,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "SatelliteException.h"
+#include "csgp4/SatelliteException.h"
