+++ src/TimeSpan.cpp
@@ -15,4 +15,4 @@
  */
 
 
-#include "TimeSpan.h"
+#include "csgp4/TimeSpan.h"
