+++ src/Util.cpp
@@ -16,7 +16,7 @@
  */
 
 
-#include "Util.h"
+#include "csgp4/Util.h"
 
 #include <algorithm>
 #include <locale>
