+++ src/csgp4/Globals.h
