+++ src/csgp4/SatelliteException.h
