+++ src/csgp4/SolarPosition.h
