+++ src/csgp4/TimeSpan.h
