+++ src/csgp4/config.h
@@ -1,0 +1,12 @@
+
+
+
+
+#ifndef CONFIG_H_IN
+#define CONFIG_H_IN
+
+#define GIT_TAG "91a95be"
+#define LIBCSGP4_VERSION "0.1.0"
+
+#endif
+
